Doc finale Aurélien
Derniers changements du code :

~ fenetre de résultat non modal -> ca bloquait le serveur !!! du coup toutes les créatures mourraient chez les clients

- les cheats

~ amélioration du panel de lancement de créatures (meilleure màj)

~ les pieces d'or sont maintenant en double -> on est obligé car les revenus sont en double... c'est nettement plus précis!

+++ trunk/src/reseau/jeu/client/ClientJeu.java
@@ -652,7 +652,7 @@ public class ClientJeu implements ConstantesServeurJeu, Runnable{
         
         log("Réception de l'état d'un joueur (id:"+idJoueur+")");
         
-        int nbPiecesDOr = message.getInt("NB_PIECES_OR");
+        double nbPiecesDOr = message.getDouble("NB_PIECES_OR");
         int score = message.getInt("SCORE");
         double revenu = message.getDouble("REVENU");
         int nbViesRestantes = message.getInt("NB_VIES_RESTANTES_EQUIPE");
